+++ backend/integration_tests/src/communities/mod.rs
@@ -4,6 +4,7 @@ mod convert_group_into_community_tests;
 mod create_channel_tests;
 mod delete_channel_tests;
 mod delete_community_tests;
+mod delete_history_tests;
 mod disappearing_message_tests;
 mod import_group_tests;
 mod join_channel_tests;