+++ backend/canisters/local_user_index/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Added
+
+- Added `HistoryDeleted` chat event variant ([#8855](https://github.com/open-chat-labs/open-chat/pull/8855))
+
 ## [[2.0.1948](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1948-local_user_index)] - 2026-01-14
 
 ### Added